batch: Move presence claims into claim helpers
The ownership package now has a separate module for baseline references and
a preparatory boundary test for presence claim records.

PresenceClaim still sits on the aggregate ownership module while
ownership_claims builds and formats the same record. That keeps claim
construction dependent on the aggregate module and preserves a cycle inside
the batch ownership package.

This commit moves PresenceClaim into batch.ownership_claims, imports it
into ownership.py under a private alias, and updates detached ownership
copying to use the claim helper module directly.

The final tests commit will remove the migration fallback from the
presence-record architecture guard.
