+++ src/git_stage_batch/batch/ownership_claims.py
@@ -4,13 +4,9 @@ from __future__ import annotations
 
 from collections.abc import Iterable
 from dataclasses import dataclass, field
-from typing import TYPE_CHECKING
 
 from ..core.line_selection import LineRanges, LineSelection
-
-if TYPE_CHECKING:
-    from .ownership import PresenceClaim
-    from .ownership_references import BaselineReference
+from .ownership_references import BaselineReference
 
 
 def parse_ownership_line_ranges(line_ranges: list[str] | list[int]) -> LineRanges:
@@ -62,13 +58,58 @@ class LineRangeBuilder:
         return LineRanges.from_ranges(ranges)
 
 
+@dataclass
+class PresenceClaim:
+    """A presence constraint over batch-source lines.
+
+    Presence claims are the first-class representation for content that must
+    exist after a batch is applied. Source lines identify the content in the
+    batch source; optional baseline references record where those source lines came
+    from in the original index/tree diff.
+    """
+
+    source_lines: list[str]
+    baseline_references: dict[int, BaselineReference] = field(default_factory=dict)
+
+    def source_line_set(self) -> LineRanges:
+        """Return batch-source line numbers covered by this presence claim."""
+        return parse_ownership_line_ranges(self.source_lines)
+
+    def to_dict(self) -> dict:
+        """Serialize to metadata dictionary."""
+        data = {"source_lines": self.source_lines}
+        if self.baseline_references:
+            data["baseline_references"] = {
+                str(line): reference.to_dict()
+                for line, reference in sorted(self.baseline_references.items())
+            }
+        return data
+
+    @classmethod
+    def from_dict(
+        cls,
+        data: dict,
+        blob_contents: dict[str, bytes] | None = None,
+    ) -> PresenceClaim:
+        """Deserialize from metadata dictionary."""
+        references_metadata = data.get("baseline_references", {})
+        return cls(
+            source_lines=data.get("source_lines", []),
+            baseline_references={
+                int(line): BaselineReference.from_dict(
+                    reference,
+                    blob_contents,
+                )
+                for line, reference in references_metadata.items()
+            },
+        )
+
+
 def presence_claims_from_source_lines(
     source_lines: LineSelection | Iterable[int],
     baseline_references: dict[int, BaselineReference] | None = None,
 ) -> list[PresenceClaim]:
     """Build normalized presence claims from a source-line selection."""
-    from .ownership import PresenceClaim
-
     source_selection = (
         source_lines
         if isinstance(source_lines, LineRanges)
