+++ src/git_stage_batch/batch/ownership_detachment.py
@@ -4,8 +4,9 @@ from __future__ import annotations
 
 from ..core.buffer import LineBuffer
 from .absence_content import copy_absence_content as _copy_absence_content
-from .ownership import AbsenceClaim, BatchOwnership, PresenceClaim, ReplacementUnit
+from .ownership import AbsenceClaim, BatchOwnership, ReplacementUnit
 from .ownership_acquisition import AcquiredBatchOwnership
+from .ownership_claims import PresenceClaim
 
 
 def acquire_detached_batch_ownership(
